Renamed file to snp_bt_ros_nodes
