+++ snp_application/include/bt/snp_bt_ros_nodes.h
